libquest: Fix alternative location quest loading
With the hack2 default episode and QuestSets, the alternative location
quests were loaded but were not placed correctly in the Registry.

The alternative location quests are loaded as "quests.QuestClassName",
so the "episode" calculated for these quests is always "quests".

This patch loads this kind of quests on any episode.

We also need to call the `load` for this alternative location before the
episode load, because the QuestSets are created on the "hack2" quest and
quests are added to the QuestSets during the creation, so the custom
quests should be loaded before.

This patch also take care of the --set-quest=QuestClassName parameter to
launch the quest correctly without the need of an episode_name.

https://phabricator.endlessm.com/T31193